+++ src/modules/rules-runner.ts
@@ -1,16 +1,19 @@
 import { config } from "../../package.json";
 import { getString } from "../utils/locale";
 import { logError } from "../utils/logger";
-import { waitUtilAsync } from "../utils/wait";
+import { timeoutPromise, waitUtilAsync } from "../utils/wait";
 import { RuleBase } from "./rules/rule-base";
+import PQueue from "p-queue";
 
 export class LintRunner {
     items: Zotero.Item[];
     rules: RuleBase<any>[];
     popWin: any;
+    // queue;
     constructor(items: Zotero.Item[], rules: RuleBase<any> | RuleBase<any>[]) {
         this.items = items;
         this.rules = Array.isArray(rules) ? rules.flat() : [rules];
+        // this.queue = new PQueue({ concurrency: 10, autoStart: false, timeout: 9000, throwOnTimeout: true });
     }
 
     async run(item: Zotero.Item) {
@@ -19,6 +22,7 @@ export class LintRunner {
             item = await rule.apply(item);
         }
         await item.saveTx();
+        // await Zotero.Promise.delay(1000);
     }
 
     async runInBatch(): Promise<void> {
@@ -70,8 +74,12 @@ export class LintRunner {
 
         for (const item of this.items) {
             if (!state) break;
-            await this.run(item)
+            await timeoutPromise(this.run(item), 9000)
                 .then(() => {
+                    if (item.getTags().includes({ tag: "linter/error", type: 1 })) {
+                        item.removeTag("linter/error");
+                        item.saveTx();
+                    }
                     current++;
                     progress.changeLine({
                         text: `[${current}/${total}] ${getString("info-batch-running")}`,
@@ -85,18 +93,65 @@ export class LintRunner {
                         text: `${getString("info-batch-has-error")}, ${err}`,
                     });
                     logError(err, item);
+                    item.setTags(["linter/error"]);
+                    item.saveTx();
                     errNum++;
                 });
         }
 
-        progress.changeLine({
-            // type: "default",
-            text: `[✔️${current} ${errNum ? ", ❌" + errNum : ""}] ${getString("info-batch-finish")}`,
-            progress: 100,
-            idx: 0,
-        });
-        progress.startCloseTimer(5000);
+        progress
+            .changeLine({
+                // type: "default",
+                text: `[✔️${current} ${errNum ? ", ❌" + errNum : ""}] ${getString("info-batch-finish")}`,
+                progress: 100,
+                idx: 0,
+            })
+            .changeLine({ text: "Finished", idx: 1 })
+            .startCloseTimer(5000);
         const endTime = new Date();
         ztoolkit.log(`[Runner] The batch tasks done in ${(endTime.getTime() - startTime.getTime()) / 1000}s`);
+
+        // const tasks = this.items.map((item) => {
+        //     return async () => {
+        //         if (!state) {
+        //             this.queue.clear();
+        //         }
+        //         await this.run(item);
+        //     };
+        // });
+
+        // this.queue.addAll(tasks);
+
+        // this.queue
+        //     .on("completed", (result) => {
+        //         current++;
+        //         progress.changeLine({
+        //             text: `[${current}/${total}] ${getString("info-batch-running")}`,
+        //             progress: (current / total) * 100,
+        //             idx: 0,
+        //         });
+        //     })
+        //     .on("error", (error) => {
+        //         progress.createLine({
+        //             type: "fail",
+        //             text: `${getString("info-batch-has-error")}, ${error}`,
+        //         });
+        //         logError(error);
+        //         errNum++;
+        //     })
+        //     .on("idle", () => {
+        //         progress
+        //             .changeLine({
+        //                 // type: "default",
+        //                 text: `[✔️${current} ${errNum ? ", ❌" + errNum : ""}] ${getString("info-batch-finish")}`,
+        //                 progress: 100,
+        //                 idx: 0,
+        //             })
+        //             .changeLine({ text: "Finished", idx: 1 });
+        //         progress.startCloseTimer(5000);
+        //         const endTime = new Date();
+        //         ztoolkit.log(`[Runner] The batch tasks done in ${(endTime.getTime() - startTime.getTime()) / 1000}s`);
+        //     });
+        // this.queue.start();
     }
 }